F705 - Lab_Sim/ClassCod/Monte_Carlo/requirements.txt
pypi
typing
pydantic

numpy
matplotlib
pyyaml